core functionality is there just need to make it look purty

+++ app/controllers/admin_controller.rb
@@ -94,9 +94,12 @@ class AdminController < ApplicationController
 	def delete
 		self.isadmin
 		eid = params[:eid]
-		Event.find(eid).destroy
-		Answer.where(eid: eid).destroy_all
-		Question.where(eid: eid).destroy_all
+		if Event.exists? id: eid
+			Event.find(eid).destroy
+			Answer.where(eid: eid).destroy_all
+			Question.where(eid: eid).destroy_all
+			Inventorie.where(eid: eid).destroy_all
+		end
 		redirect_to '/admin'
 	end
